fix(tests): resolve Pinia storeToRefs crash and stabilize unit tests
Implements safeStoreToRefs utility to handle null values in Pinia stores during testing. Updates AppState and composables to use this safe utility. Enhances Vue mocks to support toRef/toRefs. Fixes all unit tests.
