+++ __tests__/composables/useUserState-jitter.test.js
@@ -35,8 +35,48 @@ jest.unstable_mockModule('vue', () => ({
     watch: mockWatch,
     nextTick: mockNextTick,
     markRaw: mockMarkRaw,
-    reactive: (o) => o,
-    computed: (fn) => ({ value: fn() })
+    reactive: (o) => {
+      // Simple proxy to handle ref unwrapping for Pinia
+      return new Proxy(o, {
+        get(target, prop, receiver) {
+          const val = target[prop];
+          return (val && val.__v_isRef) ? val.value : val;
+        },
+        set(target, prop, value, receiver) {
+          const current = target[prop];
+          if (current && current.__v_isRef) {
+            current.value = value;
+            return true;
+          }
+          target[prop] = value;
+          return true;
+        }
+      });
+    },
+    computed: (fn) => ({ value: fn() }),
+    effectScope: () => ({ active: true, run: fn => fn(), stop: () => {} }),
+    getCurrentScope: () => null,
+    onScopeDispose: () => {},
+    toRaw: (o) => o,
+    isRef: (r) => r?.__v_isRef === true,
+    toRef: (o, k) => ({ get value() { return o[k]; }, set value(v) { o[k] = v; }, __v_isRef: true }),
+    toRefs: (o) => {
+      const ret = {};
+      for (const k in o) ret[k] = { get value() { return o[k]; }, set value(v) { o[k] = v; }, __v_isRef: true };
+      return ret;
+    },
+    inject: () => {},
+    provide: () => {},
+    getCurrentInstance: () => null,
+    hasInjectionContext: () => false,
+    isReactive: () => false,
+    shallowRef: (v) => ({ value: v, __v_isRef: true }),
+    unref: (r) => r?.__v_isRef ? r.value : r,
+    triggerRef: () => {},
+    customRef: (factory) => {
+      const { get, set } = factory(() => {}, () => {});
+      return { get value() { return get(); }, set value(v) { set(v); }, __v_isRef: true };
+    },
 }));
 
 // Import Vue (mocked)
@@ -98,6 +138,8 @@ jest.unstable_mockModule('../../app/audio/buffer-queue-node', () => ({
 }));
 
 // Import the composable
+const { createPinia, setActivePinia } = await import('pinia');
+
 const { useUserState } = await import('../../app/composables/useUserState.js');
 
 describe('useUserState Jitter Buffer Calculation', () => {
@@ -107,6 +149,7 @@ describe('useUserState Jitter Buffer Calculation', () => {
   let dataPingCallback;
 
   beforeEach(() => {
+    setActivePinia(createPinia());
     // Reset watchers
     watchers.clear();
